refactor(cymbal): move frames (Frame/RawFrame/Context) into core/types
Relocate the frame data module (Frame, RawFrame, Context, ContextLine,
ReleaseInfo, plus the resolved-frame/release DB records) from core/frames
to core/types/frames, so the network frame data lives alongside the other
wire types in core/types. crate::frames still resolves via the compat shim.

RawFrame still references the per-language raw structs (which remain under
symbolication/langs for now) — that residual edge is removed in the next
step when the structs themselves move into core/types. Compiles (one
crate); cargo check + clippy green; resolve test passes.
